chore(embedded/store): non-thread safe ongoing tx
