+++ embedded/store/immustore.go
@@ -159,214 +159,6 @@ type refVLog struct {
 	unlockedRef *list.Element // unlockedRef == nil <-> vLog is locked
 }
 
-type OngoingTx struct {
-	st   *ImmuStore
-	snap *Snapshot
-
-	entries      []*EntrySpec
-	entriesByKey map[[sha256.Size]byte]int
-
-	metadata *TxMetadata
-
-	closed  bool
-	rwmutex sync.RWMutex
-}
-
-type EntrySpec struct {
-	Key      []byte
-	Metadata *KVMetadata
-	Value    []byte
-}
-
-func (tx *OngoingTx) WithMetadata(md *TxMetadata) *OngoingTx {
-	tx.metadata = md
-	return nil
-}
-
-func (tx *OngoingTx) IsWriteOnly() bool {
-	return tx.snap == nil
-}
-
-func (tx *OngoingTx) Set(key []byte, md *KVMetadata, value []byte) error {
-	tx.rwmutex.Lock()
-	defer tx.rwmutex.Unlock()
-
-	if tx.closed {
-		return ErrAlreadyClosed
-	}
-
-	if len(key) == 0 {
-		return ErrIllegalArguments
-	}
-
-	if len(key) > tx.st.maxKeyLen {
-		return ErrorMaxKeyLenExceeded
-	}
-
-	if len(value) > tx.st.maxValueLen {
-		return ErrorMaxValueLenExceeded
-	}
-
-	kid := sha256.Sum256(key)
-	keyRef, isKeyUpdate := tx.entriesByKey[kid]
-
-	if !isKeyUpdate && len(tx.entries) > tx.st.maxTxEntries {
-		return ErrorMaxTxEntriesLimitExceeded
-	}
-
-	if !tx.IsWriteOnly() {
-		// vLen=0 + vOff=0 + vHash=0 + txmdLen=0 + kvmdLen=0
-		var indexedValue [lszSize + offsetSize + sha256.Size + sszSize + sszSize]byte
-
-		err := tx.snap.set(key, indexedValue[:])
-		if err != nil {
-			return err
-		}
-	}
-
-	e := &EntrySpec{
-		Key:      key,
-		Metadata: md,
-		Value:    value,
-	}
-
-	if isKeyUpdate {
-		tx.entries[keyRef] = e
-	} else {
-		tx.entries = append(tx.entries, e)
-		tx.entriesByKey[kid] = len(tx.entriesByKey)
-	}
-
-	return nil
-}
-
-func (tx *OngoingTx) ExistKeyWith(prefix, neq []byte) (bool, error) {
-	return tx.snap.snap.ExistKeyWith(prefix, neq)
-}
-
-func (tx *OngoingTx) Delete(key []byte) error {
-	valRef, err := tx.Get(key)
-	if err != nil {
-		return err
-	}
-
-	if valRef.KVMetadata() != nil && valRef.KVMetadata().Deleted() {
-		return ErrKeyNotFound
-	}
-
-	return tx.Set(key, NewKVMetadata().AsDeleted(true), nil)
-}
-
-func (tx *OngoingTx) Get(key []byte) (ValueRef, error) {
-	return tx.GetWith(key, IgnoreDeleted)
-}
-
-func (tx *OngoingTx) GetWith(key []byte, filters ...FilterFn) (ValueRef, error) {
-	tx.rwmutex.RLock()
-	defer tx.rwmutex.RUnlock()
-
-	if tx.closed {
-		return nil, ErrAlreadyClosed
-	}
-
-	if tx.IsWriteOnly() {
-		return nil, ErrWriteOnlyTx
-	}
-
-	return tx.snap.GetWith(key, filters...)
-}
-
-func (tx *OngoingTx) NewKeyReader(spec *KeyReaderSpec) (*KeyReader, error) {
-	tx.rwmutex.RLock()
-	defer tx.rwmutex.RUnlock()
-
-	if tx.closed {
-		return nil, ErrAlreadyClosed
-	}
-
-	if tx.IsWriteOnly() {
-		return nil, ErrWriteOnlyTx
-	}
-
-	return tx.snap.NewKeyReader(spec)
-}
-
-func (tx *OngoingTx) Commit() (*TxHeader, error) {
-	return tx.commit(true)
-}
-
-func (tx *OngoingTx) AsyncCommit() (*TxHeader, error) {
-	return tx.commit(false)
-}
-
-func (tx *OngoingTx) commit(waitForIndexing bool) (*TxHeader, error) {
-	tx.rwmutex.Lock()
-	defer tx.rwmutex.Unlock()
-
-	if tx.closed {
-		return nil, ErrAlreadyClosed
-	}
-
-	tx.closed = true
-
-	if !tx.IsWriteOnly() {
-		err := tx.snap.Close()
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return tx.st.commit(tx, nil, waitForIndexing)
-}
-
-func (tx *OngoingTx) Cancel() error {
-	tx.rwmutex.Lock()
-	defer tx.rwmutex.Unlock()
-
-	if tx.closed {
-		return ErrAlreadyClosed
-	}
-
-	tx.closed = true
-
-	if !tx.IsWriteOnly() {
-		return tx.snap.Close()
-	}
-
-	return nil
-}
-
-func (kv *EntrySpec) Digest() [sha256.Size]byte {
-	var mdbs []byte
-
-	if kv.Metadata != nil {
-		mdbs = kv.Metadata.Bytes()
-	}
-
-	mdLen := len(mdbs)
-
-	b := make([]byte, sszSize+mdLen+len(kv.Key)+sha256.Size)
-	i := 0
-
-	if mdLen > 0 {
-		// md is only written if present for backward-compatibility
-		binary.BigEndian.PutUint16(b[i:], uint16(mdLen))
-		i += sszSize
-
-		copy(b[i:], mdbs)
-		i += mdLen
-	}
-
-	copy(b[i:], kv.Key)
-	i += len(kv.Key)
-
-	hvalue := sha256.Sum256(kv.Value)
-	copy(b[i:], hvalue[:])
-	i += sha256.Size
-
-	return sha256.Sum256(b[:i])
-}
-
 func Open(path string, opts *Options) (*ImmuStore, error) {
 	if !validOptions(opts) {
 		return nil, ErrIllegalArguments
@@ -1056,80 +848,11 @@ func (s *ImmuStore) appendData(entries []*EntrySpec, donec chan<- appendableResu
 }
 
 func (s *ImmuStore) NewWriteOnlyTx() (*OngoingTx, error) {
-	return &OngoingTx{
-		st:           s,
-		entriesByKey: make(map[[sha256.Size]byte]int),
-	}, nil
+	return newWriteOnlyTx(s)
 }
 
 func (s *ImmuStore) NewTx() (*OngoingTx, error) {
-	tx := &OngoingTx{
-		st:           s,
-		entriesByKey: make(map[[sha256.Size]byte]int),
-	}
-
-	err := s.WaitForIndexingUpto(s.committedTxID, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	tx.snap, err = s.SnapshotSince(s.committedTxID)
-	if err != nil {
-		return nil, err
-	}
-
-	tx.snap.refInterceptor = func(key []byte, valRef ValueRef) ValueRef {
-		keyRef, ok := tx.entriesByKey[sha256.Sum256(key)]
-		if !ok {
-			return valRef
-		}
-
-		entrySpec := tx.entries[keyRef]
-
-		return &ongoingValRef{
-			hc:    valRef.HC(),
-			value: entrySpec.Value,
-			txmd:  tx.metadata,
-			kvmd:  entrySpec.Metadata,
-		}
-	}
-
-	return tx, nil
-}
-
-type ongoingValRef struct {
-	value []byte
-	hc    uint64
-	txmd  *TxMetadata
-	kvmd  *KVMetadata
-}
-
-func (oref *ongoingValRef) Resolve() (val []byte, err error) {
-	return oref.value, nil
-}
-
-func (oref *ongoingValRef) Tx() uint64 {
-	return 0
-}
-
-func (oref *ongoingValRef) HC() uint64 {
-	return oref.hc
-}
-
-func (oref *ongoingValRef) TxMetadata() *TxMetadata {
-	return oref.txmd
-}
-
-func (oref *ongoingValRef) KVMetadata() *KVMetadata {
-	return oref.kvmd
-}
-
-func (oref *ongoingValRef) HVal() [sha256.Size]byte {
-	return sha256.Sum256(oref.value)
-}
-
-func (oref *ongoingValRef) Len() uint32 {
-	return uint32(len(oref.value))
+	return newReadWriteTx(s)
 }
 
 func (s *ImmuStore) commit(otx *OngoingTx, expectedHeader *TxHeader, waitForIndexing bool) (*TxHeader, error) {
