+++ embedded/store/key_reader.go
@@ -89,6 +89,10 @@ func (s *Snapshot) GetWith(key []byte, filters ...FilterFn) (valRef ValueRef, er
 	return valRef, nil
 }
 
+func (s *Snapshot) ExistKeyWith(prefix []byte, neq []byte) (bool, error) {
+	return s.snap.ExistKeyWith(prefix, neq)
+}
+
 func (s *Snapshot) History(key []byte, offset uint64, descOrder bool, limit int) (tss []uint64, err error) {
 	return s.snap.History(key, offset, descOrder, limit)
 }
